Change to collision detection, now physcomponents do collision with hitboxcomponents. This decouples simplephyscomponents from hitboxes Add remainingTime to timerComponent. we keep the initial time in the time field so it can be reset to initial value if need be instead of destroying and recreating the component Members can now be part of up to 4 groups

+++ source/components/obj.c
@@ -11,7 +11,7 @@ ObjComponent* addComponentObj(s16 entId, u16 flags, int posSourceCompType) {
     return objComp;
 }
 
-void removeComponentObj(s16 entId) {
+void removeComponentObj(int entId) {
     ObjComponent* o = &gObjCompsDense[gCompSetSparse[COMP_OBJ][entId]];
     OBJ_ATTR* thisObjBufferPtr = o->obj;
     for (int i = 0; i < gNumSpritesAllocated; i++) {
